#2 Multiple reducer
1)Объединение reducer'ов
2)Подключение инструментов разработчика (redux dev tools).
