+++ src/Actions/actions.js
@@ -1,2 +1,5 @@
 export const ADD_CASH = "ADD_CASH";
 export const GET_CASH = "GET_CASH";
+
+export const ADD_CUSTOMER = "ADD_CUSTOMER";
+export const GET_CUSTOMER = "GET_CUSTOMER";
