+++ src/App.jsx
@@ -4,7 +4,7 @@ import {ADD_CASH, GET_CASH} from "./Actions/actions";
 
 function App() {
     const dispatch = useDispatch();
-    const cash = useSelector(state => state.cash);
+    const cash = useSelector(state => state.cash.cash);
     console.log(cash);
 
     const addCash = (cash) => dispatch({
